workers/logs/requirements.txt
aiohttp==3.8.1
python-socketio-client==1.1
websocket-client==1.3.1
python-socketio==5.5.2
python-socketio==5.5.2
websockets==10.2
asyncio==3.4.3